upgrade ansible on rh6

+++ bootstrap.sh
@@ -37,8 +37,12 @@ bootstrap() {
         sudo yum install -y ansible
         major_release=$(cat /etc/redhat-release | tr -dc '0-9.'|cut -d \. -f1)
         if [ $major_release = "6" ] ; then
-          # needed for galaxy usage with python 2.6 on CentOS 6
-          sudo yum install -y python-urllib3 pyOpenSSL python2-ndg_httpsclient python-pyasn1
+          if [ -f /etc/centos-release ] ; then
+            # needed for galaxy usage with python 2.6 on CentOS 6
+            sudo yum install -y python-urllib3 pyOpenSSL python2-ndg_httpsclient python-pyasn1
+          else
+            upgrade_redhat_6
+          fi
         fi
     elif [ `uname -s` = "Darwin" ] ; then
         echo "Install Homebrew packages ..."
